Finsihed redis cache code and added logging to tsdatautils

+++ packages/redis-cache/src/data/basic-redis-cache.ts
@@ -1,4 +1,4 @@
-import { IBasicCache, IBasicTimedCache, IClearableCache, Dictionary, IBasicAsyncCache, IBasicAsyncTimedCache, IAsyncClearableCache } from 'tsdatautils-core';
+import { IBasicCache, IBasicTimedCache, IClearableCache, Dictionary, IBasicAsyncCache, IBasicAsyncTimedCache, IAsyncClearableCache, ILogger } from 'tsdatautils-core';
 import * as moment from 'moment';
 import { RedisClient, ClientOpts, createClient, RetryStrategyOptions, RetryStrategy } from 'redis';
 import { ClientResponse } from 'http';
@@ -6,27 +6,54 @@ import { promisify } from 'util';
 
 export class BasicRedisCache implements IBasicAsyncCache, IBasicAsyncTimedCache, IAsyncClearableCache {
     private static clients: Dictionary<RedisClient>;
-    private clientKey: string = null;
 
-    public constructor(clientKey: string, port: number = 6379, hostIp: string = '127.0.0.1', authPass: string = null, options: ClientOpts = {}) {
+    public logger: ILogger = null;
+    public waitTimeCommandSeconds: number = 5;
+
+    private clientKey: string = null;
+    private port: number = 6379;
+    private hostIp: string = '127.0.0.1';
+    private authPass: string = null;
+    private options: ClientOpts = {};
+
+    public constructor(logger: ILogger = null, clientKey: string, hostIp: string = '127.0.0.1', port: number = 6379, authPass: string = null, options: ClientOpts = {}) {
         if (clientKey === undefined || clientKey === null) {
             throw new Error('Cannot have a null or undefined client key');
         }
+        
         this.clientKey = clientKey;
+        this.port = port;
+        this.hostIp = hostIp;
+        this.authPass = authPass;
+        this.options = options;
+
+        this.logger = logger;
         
         if (BasicRedisCache.clients === undefined) {
             BasicRedisCache.clients = {};
         }
         
         if (BasicRedisCache.clients[clientKey] === undefined) {
-            BasicRedisCache.createClient(clientKey, port, hostIp, authPass, options);
+            BasicRedisCache.createClient(clientKey, port, hostIp, authPass, options, logger);
         }
     }
 
-    private static createClient(clientKey: string, port: number = 6379, hostIp: string = '127.0.0.1', authPass: string = null, options: ClientOpts = {}): void {
+    // tslint:disable-next-line: function-name
+    public static forceCloseAllClients(): void {
+        for (let key of Object.keys(BasicRedisCache.clients)) {
+            let currentClient: RedisClient = BasicRedisCache.clients[key];
+            currentClient.end(true);
+        }
+    }
+
+    private static createClient(clientKey: string, port: number = 6379, hostIp: string = '127.0.0.1', authPass: string = null, options: ClientOpts = {}, logger: ILogger = null): void {
         BasicRedisCache.clients[clientKey] = null;
         try {
-            options.password = authPass;
+            
+            if (authPass !== undefined && authPass !== null) {
+                options.password = authPass;
+            }
+            
             options.port = port;
             options.host = hostIp;
 
@@ -52,37 +79,71 @@ export class BasicRedisCache implements IBasicAsyncCache, IBasicAsyncTimedCache,
 
             let redisClient: RedisClient = createClient(options);
 
-            //BasicRedisCache.registerClientEventHandlers(redisClient);
+            BasicRedisCache.registerClientEventHandlers(clientKey, redisClient, logger);
             BasicRedisCache.clients[clientKey] = redisClient;
         } catch (ex) {
-            throw ex;
-        } finally {
             BasicRedisCache.clients[clientKey] = undefined;
+            throw ex;
         }
     }
 
-    //private static registerClientEventHandlers(redisClient: RedisClient): void {
+    private static registerClientEventHandlers(clientKey: string, redisClient: RedisClient, logger: ILogger, options: ClientOpts = {}): void {
+        redisClient.on('connect', () => { 
+            if (logger !== undefined && logger !== null) {
+                logger.logInfo('Redis ' + clientKey + ' connected.');
+            }
+        });
 
-    //}
+        redisClient.on('ready', () => { 
+            if (logger !== undefined && logger !== null) {
+                logger.logInfo('Redis ' + clientKey + ' ready.');
+            }
+        });
+
+        redisClient.on('reconnecting', () => { 
+            if (logger !== undefined && logger !== null) {
+                logger.logInfo('Redis ' + clientKey + ' reconnecting.');
+            }
+        });
+
+        redisClient.on('end', () => { 
+            if (logger !== undefined && logger !== null) {
+                logger.logInfo('Redis ' + clientKey + ' ended.');
+            }
+        });
+
+        redisClient.on('error', (err: any) => {
+            BasicRedisCache.handleError(err, logger, clientKey, options);
+        });
+    }
+
+    private static handleError(err: any, logger: ILogger, clientKey: string, options: ClientOpts = {}): void {
+        logger.logError(err);
+
+        BasicRedisCache.createClient(clientKey, options.port, options.host, options.password, options, logger);
+    }
 
     public getItemAsync<T>(key: string): Promise<T> {
         return new Promise<T>((resolve: (val: T) => void, reject: (reason: any) => void) => {
             try {
                 let client: RedisClient = BasicRedisCache.clients[this.clientKey];
                 this.throwIfClientDoesNotExist(client);
-                this.throwIfNotConnected(client);
-
-                let getAsync: any = promisify(client.get).bind(client);
-                getAsync(key).then((res: string) => {
-                    try {
-                        let resParsed: T = JSON.parse(res);
-                        resolve(resParsed);
-                    } catch (err) {
+                this.confirmConnection(client).then(() => {
+                    let getAsync: any = promisify(client.get).bind(client);
+                    // tslint:disable-next-line: no-unsafe-any
+                    getAsync(key).then((res: string) => {
+                        try {
+                            let resParsed: T = JSON.parse(res);
+                            resolve(resParsed);
+                        } catch (err) {
+                            reject(err);
+                        }
+                    }).catch((err: any) => {
                         reject(err);
-                    }
-                }).catch((err: any) => {
+                    });
+                }).catch((err: string) => {
                     reject(err);
-                });
+                });                
             } catch (err) {
                 reject(err);
             }
@@ -94,24 +155,26 @@ export class BasicRedisCache implements IBasicAsyncCache, IBasicAsyncTimedCache,
             try {
                 let client: RedisClient = BasicRedisCache.clients[this.clientKey];
                 this.throwIfClientDoesNotExist(client);
-                this.throwIfNotConnected(client);
+                this.confirmConnection(client).then(() => {
+                    let itemStringified: string = JSON.stringify(item);
 
-                let itemStringified: string = JSON.stringify(item);
+                    let setAsync: any = promisify(client.set).bind(client);
+                    let setPromise: any = null;
 
-                let setAsync: any = promisify(client.set).bind(client);
-                let setPromise: any = null;
+                    if (ttl === null) {
+                        setPromise = setAsync(key, itemStringified);
+                    } else {
+                        setPromise = setAsync(key, itemStringified, 'EX', ttl.asSeconds());
+                    }
 
-                if (ttl === null) {
-                    setPromise = setAsync(key, itemStringified);
-                } else {
-                    setPromise = setAsync(key, itemStringified, 'EX', ttl.asSeconds());
-                }
-
-                setPromise.then((res: string) => {
-                    resolve(true);
-                }).catch((err: any) => {
+                    setPromise.then((res: string) => {
+                        resolve(true);
+                    }).catch((err: any) => {
+                        reject(err);
+                    });
+                }).catch((err: string) => {
                     reject(err);
-                });
+                });                
             } catch (err) {
                 reject(err);
             }
@@ -123,15 +186,17 @@ export class BasicRedisCache implements IBasicAsyncCache, IBasicAsyncTimedCache,
             try {
                 let client: RedisClient = BasicRedisCache.clients[this.clientKey];
                 this.throwIfClientDoesNotExist(client);
-                this.throwIfNotConnected(client);
+                this.confirmConnection(client).then(() => {
+                    let delAsync: any = promisify(client.del).bind(client);
 
-                let delAsync: any = promisify(client.del).bind(client);
-
-                delAsync(key).then((res: number) => {
-                    resolve(true);
-                }).catch((err: any) => {
+                    delAsync(key).then((res: number) => {
+                        resolve(true);
+                    }).catch((err: any) => {
+                        reject(err);
+                    });
+                }).catch((err: string) => {
                     reject(err);
-                });
+                });                
             } catch (err) {
                 reject(err);
             }
@@ -143,13 +208,15 @@ export class BasicRedisCache implements IBasicAsyncCache, IBasicAsyncTimedCache,
             try {
                 let client: RedisClient = BasicRedisCache.clients[this.clientKey];
                 this.throwIfClientDoesNotExist(client);
-                this.throwIfNotConnected(client);
+                this.confirmConnection(client).then(() => {
+                    let flushAllAsync: any = promisify(client.flushall).bind(client);
 
-                let flushAllAsync: any = promisify(client.flushall).bind(client);
-
-                flushAllAsync().then(() => {
-                    resolve();
-                }).catch((err: any) => {
+                    flushAllAsync().then(() => {
+                        resolve();
+                    }).catch((err: any) => {
+                        reject(err);
+                    });
+                }).catch((err: string) => {
                     reject(err);
                 });
             } catch (err) {
@@ -164,9 +231,31 @@ export class BasicRedisCache implements IBasicAsyncCache, IBasicAsyncTimedCache,
         }
     }
 
-    private throwIfNotConnected(client: RedisClient) {
-        if (!client.connected) {
-            throw new Error('Redis not connected');
+
+
+    private confirmConnection(client: RedisClient): Promise<void> {
+        return new Promise<void>((resolve: () => void, reject: (reason: any) => void) => {
+            if (client.connected) {
+                resolve();
+            } else {
+                let stopWaitingTime: moment.Moment = moment().add(this.waitTimeCommandSeconds, 'seconds');
+                this.waitForConnection(client, stopWaitingTime, resolve, reject);
+            }
+        });
+    }
+
+    private waitForConnection(client: RedisClient, stopWaitingTime: moment.Moment, resolve: () => void, reject: (reason: any) => void): void {
+        if (client.connected) {
+            resolve();
+        } else {
+            setTimeout(() => {
+                if (moment().isAfter(stopWaitingTime)) {
+                    reject('Not connected');
+                } else {
+                    this.waitForConnection(client, stopWaitingTime, resolve, reject);
+                }
+            // tslint:disable-next-line: align
+            }, 500);
         }
     }
